removed proof-of-concept notebooks no longer relevant to project
