+++ src/components/incose.py
@@ -6,33 +6,15 @@ build requirement revision prompt templates, and run prompts for requirement rev
 from __future__ import annotations
 
 import re
-import logging
-import asyncio
 from pathlib import Path
 from functools import partial
 from typing import List, Dict, Optional, Any, Callable
 
 import pandas as pd
-import nest_asyncio
 from pprint import pformat
-from tqdm import tqdm
-from langchain_core.runnables import (
-    RunnableLambda,
-    RunnableParallel,
-    RunnablePassthrough,
-    RunnableSequence,
-)
 
-import src
-from src.prj_logger import get_logs
-from src import utils
-from src.utils import map_A_to_B
 import src.components.prompteval as pe
-from src.components.promptrunner import PromptRunner
 from src.components.preprocess import TextPreprocessor, Sectionalize, BuildTemplates
-
-nest_asyncio.apply()
-
 
 # -------------------------------------------------------------------------
 # -------------------------- PREPROCESSING --------------------------------
